src/components: move UserSelect
