+++ src/components/global/DrawerHeader.vue
@@ -30,7 +30,7 @@ import { defineComponent, computed } from 'vue';
 
 // import components
 import HelpButton from './global/HelpButton.vue';
-import UserSelect from './UserSelect.vue';
+import UserSelect from './global/UserSelect.vue';
 
 export default defineComponent({
   name: 'DrawerHeader',
